Add files via upload
This is the program to solve the 1st problem in the test I was sent.  An example of the console output is included.  The data.txt file used to create the birth and death dates is in the Debug folder of the application and can be modified to try new data sets.